review bitops code

+++ keepalived/include/vrrp.h
@@ -96,7 +96,7 @@ typedef struct _vrrp_t {
 	vrrp_sgroup_t		*sync;			/* Sync group we belong to */
 	interface_t		*ifp;			/* Interface we belong to */
 	int			dont_track_primary;	/* If set ignores ifp faults */
-	int			vmac_flags;		/* VRRP VMAC flags */
+	unsigned long		vmac_flags;		/* VRRP VMAC flags */
 	char			vmac_ifname[IFNAMSIZ];	/* Name of VRRP VMAC interface */
 	unsigned int		vmac_ifindex;		/* ifindex of vmac interface */
 	list			track_ifp;		/* Interface state we monitor */
